feat: rework the package from scratch
fix #4
fix #2
