+++ playground/index.html
@@ -16,6 +16,7 @@
       <a href="https://plausible.io/docs/plausible-script">Documentation (same page)</a>
       <button id="navigation">Click me (navigation event)</button>
       <button id="btn">Click me (custom event)</button>
+      <button id="cleanup">Cleanup</button>
     </div>
     <script type="module" src="/src/main.ts"></script>
   </body>
