+++ src/extensions/index.ts
@@ -1,0 +1,2 @@
+export { useAutoPageviews } from './auto-pageviews'
+export { useAutoOutboundTracking } from './auto-outbound-tracking'
